Make ruff ignore E402 (Module-level import not at top of cell)
For performance reasons, only the `sys` module is imported at the top of the file. It is used to check the minimum required Python version and if the platform is supported. If any of these checks fail, the script is interrupted with an error and does not waste time importing anything that would not be used.