Improved how the installation directory is defined.
Added a flag --user that indicates the installation must be placed inside the user's home directory (~/.local). This flag is false by default.
Added an option --prefix to enable the application to be installed in a given directory instead of default locations.
Defined the default directory for installation as /usr/local. This requires that the user running the script has administrative privileges to write to that directory.